Improvements to rotational IK
Rotation in the world frame is finally starting to resemble a stable and
functional feature. It's not quite there, since we still have no
limitations on where joints are rotated to, and I noticed some odd
behavior from this system from just a cursory test, but this is a big
step forward from the previous iterations.